Fixed issue #3 , cleaner functions
Server doesn't crash anymore.
Renamed events.
'question number' event is sent to **boitier** only when teacher clicks "next question" --> Initial vlaue of the question number (on the UI) must be 1 (same as we do for phones)
Removed console.log.
Removed argument to getQuestion and getSession, we now check inside the getQuestion if we are at the last question.
Can now refresh page when displaying a question (wont go from first question to second for no reason).